:recycle: api: moved userRepository to src/shared
