+++ api/src/certification/shared/domain/usecases/index.js
@@ -3,7 +3,7 @@ import { dirname, join } from 'node:path';
 
 import * as sessionCodeService from '../../../../../lib/domain/services/session-code-service.js';
 import * as sessionValidator from '../../../../../lib/domain/validators/session-validator.js';
-import * as userRepository from '../../../../../lib/infrastructure/repositories/user-repository.js';
+import * as userRepository from '../../../../../src/shared/infrastructure/repositories/user-repository.js';
 import * as sessionRepository from '../../../session/infrastructure/repositories/session-repository.js';
 import * as certificationCenterRepository from '../../../shared/infrastructure/repositories/certification-center-repository.js';
