+++ api/src/shared/infrastructure/repositories/user-repository.js
@@ -1,29 +1,28 @@
 import moment from 'moment';
-
-import { knex } from '../../../db/knex-database-connection.js';
-import { DomainTransaction } from '../DomainTransaction.js';
-import { BookshelfUser } from '../orm-models/User.js';
-import { fetchPage, isUniqConstraintViolated } from '../utils/knex-utils.js';
-import * as bookshelfToDomainConverter from '../utils/bookshelf-to-domain-converter.js';
+import { knex } from '../../../../db/knex-database-connection.js';
+import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
+import { BookshelfUser } from '../../../../lib/infrastructure/orm-models/User.js';
+import { isUniqConstraintViolated, fetchPage } from '../../../../lib/infrastructure/utils/knex-utils.js';
+import * as bookshelfToDomainConverter from '../../../../lib/infrastructure/utils/bookshelf-to-domain-converter.js';
 
 import {
   AlreadyExistingEntityError,
   AlreadyRegisteredEmailError,
   AlreadyRegisteredUsernameError,
   UserNotFoundError,
-} from '../../domain/errors.js';
+} from '../../../../lib/domain/errors.js';
 
-import { User } from '../../domain/models/User.js';
-import { UserDetailsForAdmin } from '../../domain/models/UserDetailsForAdmin.js';
-import { Membership } from '../../domain/models/Membership.js';
-import { CertificationCenter } from '../../domain/models/CertificationCenter.js';
-import { CertificationCenterMembership } from '../../domain/models/CertificationCenterMembership.js';
-import { Organization } from '../../domain/models/Organization.js';
-import { OrganizationLearnerForAdmin } from '../../domain/read-models/OrganizationLearnerForAdmin.js';
-import { AuthenticationMethod } from '../../domain/models/AuthenticationMethod.js';
-import { NON_OIDC_IDENTITY_PROVIDERS } from '../../domain/constants/identity-providers.js';
-import * as OidcIdentityProviders from '../../domain/constants/oidc-identity-providers.js';
-import { UserLogin } from '../../domain/models/UserLogin.js';
+import { User } from '../../../../lib/domain/models/User.js';
+import { UserDetailsForAdmin } from '../../../../lib/domain/models/UserDetailsForAdmin.js';
+import { Membership } from '../../../../lib/domain/models/Membership.js';
+import { CertificationCenter } from '../../../../lib/domain/models/CertificationCenter.js';
+import { CertificationCenterMembership } from '../../../../lib/domain/models/CertificationCenterMembership.js';
+import { Organization } from '../../../../lib/domain/models/Organization.js';
+import { OrganizationLearnerForAdmin } from '../../../../lib/domain/read-models/OrganizationLearnerForAdmin.js';
+import { AuthenticationMethod } from '../../../../lib/domain/models/AuthenticationMethod.js';
+import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../lib/domain/constants/identity-providers.js';
+import * as OidcIdentityProviders from '../../../../lib/domain/constants/oidc-identity-providers.js';
+import { UserLogin } from '../../../../lib/domain/models/UserLogin.js';
 
 const getByEmail = async function (email) {
   const foundUser = await knex.from('users').whereRaw('LOWER("email") = ?', email.toLowerCase()).first();
