+++ api/tests/integration/domain/event/handle-pole-emploi-participation-started_test.js
@@ -14,7 +14,7 @@ import * as campaignParticipationRepository from '../../../../lib/infrastructure
 import * as organizationRepository from '../../../../lib/infrastructure/repositories/organization-repository.js';
 import * as poleEmploiSendingRepository from '../../../../lib/infrastructure/repositories/pole-emploi-sending-repository.js';
 import * as targetProfileRepository from '../../../../lib/infrastructure/repositories/target-profile-repository.js';
-import * as userRepository from '../../../../lib/infrastructure/repositories/user-repository.js';
+import * as userRepository from '../../../../src/shared/infrastructure/repositories/user-repository.js';
 import { CampaignParticipationStarted as CampaignParticipationStartedEvent } from '../../../../lib/domain/events/CampaignParticipationStarted.js';
 
 describe('Integration | Event | Handle Pole emploi participation started', function () {
