+++ api/tests/integration/shared/infrastructure/repositories/user-repository_test.js
@@ -1,6 +1,6 @@
 import lodash from 'lodash';
 const { each, map, times, pick } = lodash;
-import { expect, knex, databaseBuilder, catchErr, sinon } from '../../../test-helper.js';
+import { expect, knex, databaseBuilder, catchErr, sinon } from '../../../../test-helper.js';
 
 import {
   AlreadyExistingEntityError,
@@ -8,19 +8,19 @@ import {
   AlreadyRegisteredUsernameError,
   NotFoundError,
   UserNotFoundError,
-} from '../../../../lib/domain/errors.js';
+} from '../../../../../lib/domain/errors.js';
 
-import { User } from '../../../../lib/domain/models/User.js';
-import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../lib/domain/constants/identity-providers.js';
-import { UserDetailsForAdmin } from '../../../../lib/domain/models/UserDetailsForAdmin.js';
-import { Membership } from '../../../../lib/domain/models/Membership.js';
-import { CertificationCenter } from '../../../../lib/domain/models/CertificationCenter.js';
-import { CertificationCenterMembership } from '../../../../lib/domain/models/CertificationCenterMembership.js';
-import { Organization } from '../../../../lib/domain/models/Organization.js';
-import { OrganizationLearnerForAdmin } from '../../../../lib/domain/read-models/OrganizationLearnerForAdmin.js';
-import * as OidcIdentityProviders from '../../../../lib/domain/constants/oidc-identity-providers.js';
-import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
-import * as userRepository from '../../../../lib/infrastructure/repositories/user-repository.js';
+import { User } from '../../../../../lib/domain/models/User.js';
+import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../../lib/domain/constants/identity-providers.js';
+import { UserDetailsForAdmin } from '../../../../../lib/domain/models/UserDetailsForAdmin.js';
+import { Membership } from '../../../../../lib/domain/models/Membership.js';
+import { CertificationCenter } from '../../../../../lib/domain/models/CertificationCenter.js';
+import { CertificationCenterMembership } from '../../../../../lib/domain/models/CertificationCenterMembership.js';
+import { Organization } from '../../../../../lib/domain/models/Organization.js';
+import { OrganizationLearnerForAdmin } from '../../../../../lib/domain/read-models/OrganizationLearnerForAdmin.js';
+import * as OidcIdentityProviders from '../../../../../lib/domain/constants/oidc-identity-providers.js';
+import { DomainTransaction } from '../../../../../lib/infrastructure/DomainTransaction.js';
+import * as userRepository from '../../../../../src/shared/infrastructure/repositories/user-repository.js';
 
 const expectedUserDetailsForAdminAttributes = [
   'id',
